+++ src/service/collection/collection.service.ts
@@ -2,53 +2,53 @@
  * (Posts, Pages, Comments, Media, Custom Endpoint ... etc)
  * */
 
-import {Injectable} from "@angular/core";
-import {Observable} from "rxjs/Observable";
+import { Injectable } from '@angular/core';
+import { Headers } from '@angular/http';
+import { Observable } from 'rxjs/Observable';
 
-import {WpQueryArgs} from "../../helpers/wp-query.class";
-import {WpHttp} from "../../helpers/wp-http.class";
-import {CollectionInterface} from "./collection.interface";
-import {Headers} from '@angular/http';
+import { WpHttp } from '../../classes/wp-http.class';
+import { CollectionInterface, CollectionResponse } from './collection.interface';
+import { WpPagination } from '../../classes/wp-pagination.class';
 
 @Injectable()
 export class CollectionService implements CollectionInterface {
 
   /** Request Parameter  */
-  private _args: WpQueryArgs;
+  args;
 
   /** Collection Pagination Properties */
-  private _pagination: WpPagination;
-  private _items = [];
+  pagination: WpPagination;
+  items = [];
 
   constructor(private http: WpHttp, private endpoint: string) {
-    this._args = new WpQueryArgs({});
+    this.args = {};
   }
 
   /**
    * Get the collection
    * @param args
-   * @returns {Observable<Response>}
+   * @returns {Observable<Any>}
    */
-  public get = (args?: WpQueryArgs): Observable<any> => {
+  get(args?): Observable<CollectionResponse> {
     /** reset pagination */
-    this._pagination = new WpPagination();
+    this.pagination = new WpPagination();
     if (args) {
-      this._args = args;
+      this.args = args;
       /** if args.page is provided set the pagination currenPage */
       if (args.page) {
-        this._pagination.currentPage = args.page;
+        this.pagination.currentPage = args.page;
       }
     }
-    return this.fetch().map((res)=> {
+    return this.fetch().map((res) => {
 
-      if(res.error){
+      if (res.error) {
         return res;
       }
 
-      this._items = res;
+      this.items = res;
       return {
-        data: this._items,
-        pagination: this._pagination
+        data: this.items,
+        pagination: this.pagination
       };
     });
   };
@@ -57,20 +57,20 @@ export class CollectionService implements CollectionInterface {
    * Get next page of the collection combined with current collection
    * @returns {any}
    */
-  public more = (): Observable<any> => {
-    if (this._pagination && this._pagination.hasMore) {
+  more(): Observable<CollectionResponse> {
+    if (this.pagination && this.pagination.hasMore) {
       /** increment currentPage then set page argument */
-      this._args.page = ++this._pagination.currentPage;
-      return this.fetch().map((res)=> {
-      
-        if(res.error){
+      this.args.page = ++this.pagination.currentPage;
+      return this.fetch().map((res) => {
+
+        if (res.error) {
           return res;
         }
 
-        this._items = this._items.concat(res);
+        this.items = this.items.concat(res);
         return {
-          data: this._items,
-          pagination: this._pagination
+          data: this.items,
+          pagination: this.pagination
         };
       });
     }
@@ -83,20 +83,20 @@ export class CollectionService implements CollectionInterface {
    * Get next page of the collection
    * @returns {any}
    */
-  public next = (): Observable<any> => {
-    if (this._pagination && this._pagination.hasMore) {
+  next(): Observable<CollectionResponse> {
+    if (this.pagination && this.pagination.hasMore) {
       /** increment currentPage then set page argument */
-      this._args.page = ++this._pagination.currentPage;
-      return this.fetch().map((res)=> {
+      this.args.page = ++this.pagination.currentPage;
+      return this.fetch().map((res) => {
 
-        if(res.error){
+        if (res.error) {
           return res;
         }
 
-        this._items = res;
+        this.items = res;
         return {
-          data: this._items,
-          pagination: this._pagination
+          data: this.items,
+          pagination: this.pagination
         };
       });
     }
@@ -109,20 +109,20 @@ export class CollectionService implements CollectionInterface {
    * Get prev page of the collection
    * @returns {any}
    */
-  public prev = (): Observable<any> => {
-    if (this._pagination && this._pagination.hasPrev) {
+  prev(): Observable<CollectionResponse> {
+    if (this.pagination && this.pagination.hasPrev) {
       /** decrement currentPage then set page argument */
-      this._args.page = --this._pagination.currentPage;
-      return this.fetch().map((res)=> {
+      this.args.page = --this.pagination.currentPage;
+      return this.fetch().map((res) => {
 
-        if(res.error){
+        if (res.error) {
           return res;
         }
 
-        this._items = res;
+        this.items = res;
         return {
-          data: this._items,
-          pagination: this._pagination
+          data: this.items,
+          pagination: this.pagination
         };
       });
     }
@@ -135,9 +135,9 @@ export class CollectionService implements CollectionInterface {
    * Fires the final request
    * @returns {Observable<any>}
    */
-  private fetch = (): Observable<any> => {
+  private fetch(): Observable<any> {
 
-    return this.http.get(this.endpoint, this._args).map(
+    return this.http.get(this.endpoint, this.args).map(
       (res) => {
         /** Set pagination  */
         this.setPagination(res.headers);
@@ -145,7 +145,7 @@ export class CollectionService implements CollectionInterface {
       }
     ).catch((err) => {
       /** return errors in form of res.error */
-      return Observable.of({error: err});
+      return Observable.of({ error: err });
     });
   };
 
@@ -157,35 +157,16 @@ export class CollectionService implements CollectionInterface {
   private setPagination = (headers: Headers): WpPagination => {
 
     /** Fix issue of different property names in response headers */
-    this._pagination.totalPages =
-      +headers.get('X-WP-TotalPages') || +headers.get('X-Wp-TotalPages') || +headers.get('X-Wp-Totalpages') || 0;
+    this.pagination.totalPages =
+      +headers.get('x-wp-totalpages') || +headers.get('X-WP-TotalPages') || +headers.get('X-Wp-TotalPages')
+       || +headers.get('X-Wp-Totalpages') || 0;
 
-    this._pagination.totalObjects =
-      +headers.get('X-WP-Total') || +headers.get('X-Wp-Total') || 0;
+    this.pagination.totalObjects =
+      +headers.get('x-wp-total') || +headers.get('X-WP-Total') || +headers.get('X-Wp-Total') || 0;
 
-    this._pagination.links = headers.get('Link');
+    this.pagination.links = headers.get('Link') || headers.get('link');
 
-    return this._pagination;
+    return this.pagination;
   }
 
-}
-
-/**
- * Pagination class holds the current collection response pagination and links
- */
-export class WpPagination {
-  /** Pagination holds the navigation data and links provided from WP API response header*/
-  constructor(public currentPage: number = 1,
-              public totalPages: number = 0,
-              public totalObjects: number = 0,
-              public links?: any) {
-
-  }
-
-  get hasMore(): boolean {
-    return this.currentPage < this.totalPages;
-  }
-  get hasPrev(): boolean{
-    return this.currentPage > 1;
-  }
 }